+++ src/opal.js
@@ -7,11 +7,10 @@ const TargetClass = [
     String,
 ];
 
-const OriginalKeys = {};
-function restoreOriginal() {
+function restoreOriginal(originalKeys) {
     TargetClass.forEach(targetClass => {
         Object.keys(targetClass.prototype)
-            .filter(key => OriginalKeys[targetClass.name].indexOf(key) < 0)
+            .filter(key => originalKeys[targetClass.name].indexOf(key) < 0)
             .forEach(key => {
                 // eslint-disable-next-line no-param-reassign
                 delete targetClass.prototype[key];
@@ -20,13 +19,17 @@ function restoreOriginal() {
 }
 
 function backupOriginalKeys() {
+    const originalKeys = {};
+
     TargetClass.forEach(targetClass => {
-        OriginalKeys[targetClass.name] = Object.keys(targetClass.prototype);
+        originalKeys[targetClass.name] = Object.keys(targetClass.prototype);
     });
+
+    return originalKeys;
 }
 
 const OpalPrototypes = {};
-function restoreOpal() {
+function loadOpal() {
     TargetClass.forEach(targetClass => {
         Object.assign(targetClass.prototype, OpalPrototypes[targetClass.name]);
     });
@@ -37,24 +40,31 @@ function backupOpalKeys() {
     });
 }
 
-backupOriginalKeys();
-// eslint-disable-next-line
-const Opal = require('../lib/opal.ruby.js');
-backupOpalKeys();
-restoreOriginal();
+function init() {
+    const originalKeys = backupOriginalKeys();
+
+    // eslint-disable-next-line
+    const Opal = require('../lib/opal.ruby.js');
+    backupOpalKeys();
+
+    restoreOriginal(originalKeys);
+}
+init();
 
 module.exports = function opal(callback) {
+    const originalKeys = backupOriginalKeys();
+
     try {
-        restoreOpal();
+        loadOpal();
 
         // eslint-disable-next-line
         const result = callback(Opal);
 
-        restoreOriginal();
+        restoreOriginal(originalKeys);
 
         return result;
     } catch (e) {
-        restoreOriginal();
+        restoreOriginal(originalKeys);
         throw e;
     }
 };