+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+### Changed
+- Implement dynamic dicebot loader
+
 ## [0.3.2] - 2017-08-23
 ### Fixed
 - Protect another built-in objects